Completed RESTful service
Change in DbHelper: getColumnName() ignores alias defined in queries. That's why changed to getColumnLabel().

+++ gamiOffice/src/main/java/gamiOffi/components/tables/ChallengeComponent.java
@@ -47,7 +47,7 @@ public class ChallengeComponent extends Queriable {
             case componentId:
             case challengeId:
             case componentWeight:
-              pstmt.setLong(counter++, recordObject.getLong(attributeName));
+              pstmt.setDouble(counter++, recordObject.getLong(attributeName));
               break;
             case componentCode:
             case componentDesc:
